[olarft] Rename folders to saved-rows and row-editor
