+++ src/tools/larft/components/row-editor/FieldDetails.jsx
