+++ src/tools/larft/components/row-editor/ParsedRow.jsx
