+++ src/tools/larft/components/row-editor/Piped.jsx
