+++ src/tools/larft/components/row-editor/parsedHelpers.jsx
